# backend-srv/src/main/resources/openapi/preinversion/CU-PRE-03.5.openapi.yaml
openapi: 3.0.3
info:
  title: SIIP - Preinversión - CU-PRE-03.5 Selección y Registro de Etapas
  version: 1.3.0
  description: 'Fragmento OpenAPI derivado de CU-PRE-03.5 "Selección y Registro de Etapas" (UC-PRE-3_5-Seleccion_y_Registro_de_Etapas.md, versión 1.3) y sus 6 historias BDD. Actor principal: Técnico URP. Actor secundario: Coordinador SYMP (única acción: seleccionar Co-ejecutor, RN16). Introduce los recursos `RutaPreinversion`, `Etapa`, `FichaInformacionGeneral` y `FichaEmergencia`, todos hijos de `Proyecto` (CU-PRE-01.openapi.yaml, reutilizado vía `$ref` externo).

...
    en "Campos a habilitar para Actualización de O.T." (aplica / no aplica al CU / sin dato) — desbloquea la referencia que CU-PRE-26 dejaba genérica (`camposHabilitados: array of string`) por falta de esta misma confirmación. (4) Matriz B.2 (36 combinaciones) confirmada como definitiva, sin cambios de contenido adicionales.

    Quedan sin resolver, tal como antes: editabilidad de criterios de Ruta de Preinversión, nomenclatura GRD/GRC/ACC, y si "Fuentes de Financiamiento" debería tener su propia fila en el Anexo F distinta de "Presupuesto de Inversión" — ver contrato-CU-PRE-03.5.md.

...

    '
paths:
...
  /proyectos/{idProyecto}/ruta-preinversion/generacion:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      operationId: generarRutaPreinversion
      summary: Generar la Ruta de Preinversión sugerida (botón "Generar Ruta de Preinversión")
      description: 'Cálculo sin persistencia (Flujo Básico, pasos 3-4): a partir de los 3 criterios calificados, devuelve las etapas sugeridas según la tabla del Anexo B.2 (RN10). No guarda nada — la persistencia ocurre recién en `aceptarRutaPreinversion` (botón "Aceptar", FA-01) o, si el Técnico URP prefiere una ruta distinta, en `modificarRutaPreinversion` (botón "Modificar", FA-02). Solo aplica a proyectos con `iniciativaInversion = PROYECTO` (RN07/RN08); para "Estudios Generales" y "Programa" el botón "Ruta de Preinversión" está desactivado y las etapas PERFIL/EJECUCIÓN ya vienen preseleccionadas automáticamente (ver `GET /proyectos/{idProyecto}/etapas`).

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriteriosCalificacion'
      responses:
        '200':
          description: Etapas sugeridas según el Anexo B.2, sin persistir.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutaPreinversionSugerida'
        '400':
          description: 'Falta calificar alguno de los 3 criterios (RN01); escenario BDD explícito ("Intentar generar la Ruta de Preinversión sin calificar un criterio").

            '
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: El actor autenticado no es Técnico URP (inferido de `x-roles`).
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '409':
          description: 'El proyecto tiene `iniciativaInversion` distinta de PROYECTO (RN07/RN08: botón "Ruta de Preinversión" desactivado para Estudios Generales/Programa). Inferido de RN07/RN08, sin un escenario BDD que llame al endpoint en ese caso — la BDD solo describe la desactivación del botón en el cliente.

            '
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
  /proyectos/{idProyecto}/ruta-preinversion/aceptacion:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      operationId: aceptarRutaPreinversion
      summary: Aceptar la Ruta de Preinversión sugerida (botón "Aceptar", FA-01)
      description: 'Recibe los mismos 3 criterios que `generarRutaPreinversion` (el cliente ya los tiene, del paso anterior) y persiste la ruta calculada a partir de ellos, trasladando las etapas a Registro de Etapas (Anexo A.1). Se optó por recomputar en vez de requerir que el servidor recuerde una "sugerencia pendiente" entre llamadas — evita estado de sesión intermedio para una operación de dos pasos.

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriteriosCalificacion'
      responses:
        '200':
          description: Ruta de Preinversión aceptada y persistida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RutaPreinversion'
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: El actor autenticado no es Técnico URP (inferido de `x-roles`).
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
  /proyectos/{idProyecto}/etapas:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    get:
      operationId: listarEtapas
      summary: Consultar la tabla "Registro de Etapas" (Anexo A.1)
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      responses:
        '200':
          description: 'Etapas de la Ruta de Preinversión aceptada, con su estado actual. El array se devuelve siempre en el orden de la ruta (Perfil, Prefactibilidad, Factibilidad, Diseño, Ejecución) — no alfabético ni el orden de creación en base de datos.

            '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Etapa'
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
    put:
      operationId: actualizarEtapas
      summary: Registrar costo y fechas de las etapas (botón único "Guardar", FA-01.1/FA-03)
      description: 'Un solo botón "Guardar" para toda la tabla (Anexo A.1 no tiene un botón por fila); por eso este endpoint recibe y actualiza todas las etapas visibles en un solo `PUT`, no una por una. RN19 (bordes rojos por campos obligatorios incompletos — "Costo de la etapa", fechas) es puramente visual, mismo criterio que la mayoría del contrato: **no** genera 400 (corrección respecto a una versión previa de este fragmento, que trataba esta regla como rechazo real; el `.feature` actualizado confirma explícitamente "el sistema marca en rojo el borde del campo", no un rechazo de guardado). RN04: el formato de fecha dd/mm/aaaa **sí** se rechaza si no se cumple — confirmado por escenario BDD dedicado ("el sistema no acepta el valor"), distinto de RN19; el servidor también rechaza una fecha con formato válido pero que no existe en el calendario (p.ej. 31/02/2026). RN05/RN11: el `costo` de la etapa EJECUCIÓN es sobrescrito automáticamente por el servidor según el Presupuesto de inversión u Opinión Técnica más reciente — cualquier valor enviado para EJECUCIÓN en este `PUT` se ignora. RN22 sugería que
        lo mismo aplicaría a las etapas de Preinversión (costo "procederá" de CU-PRE-22.1); esto quedó confirmado al procesar ese CU (su RN06: "el valor... se debe actualizar en el campo Costo de la etapa de CU-PRE-03.5"), pero sigue sin resolver la contradicción con RN04/Anexo B.1 de este mismo CU-PRE-03.5 (que marcan `costo` editable) y con el propio camino feliz de la historia BDD de este CU (que sí registra un costo manualmente); se mantiene el comportamiento funcional de la BDD para las etapas de Preinversión — ver contrato-CU-PRE-03.5.md y contrato-CU-PRE-22.1.md. RN23 (nueva): las etapas con fechaInicio/fechaFin completos deben respetar el orden cronológico de la ruta (PERFIL/PREFACTIBILIDAD/FACTIBILIDAD/DISEÑO/EJECUCIÓN) — ninguna puede iniciar antes de que finalice la etapa previa con fechas completas, ni terminar antes de su propio inicio; se rechaza con 400, igual que RN04. Las etapas sin ambas fechas quedan fuera de esta validación (RN19).

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarEtapasRequest'
      responses:
        '200':
          description: 'Información guardada; se habilitan los botones de navegación de las etapas guardadas (FA-01.1.2/FA-03.2). El array se devuelve siempre en el orden de la ruta (Perfil, Prefactibilidad, Factibilidad, Diseño, Ejecución) — no alfabético ni el orden de creación en base de datos.

            '
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Etapa'
        '400':
          description: '`fechaInicio` o `fechaFin` no cumplen el formato dd/mm/aaaa, o son una fecha que no existe en el calendario (RN04). Escenario BDD explícito ("el sistema no acepta el valor"), sin mensaje literal documentado. A diferencia de una versión anterior de este fragmento, ya no se rechaza aquí por "Costo de la etapa"/fechas faltantes (RN19, puramente visual). También se rechaza aquí cuando las fechas de las etapas no respetan el orden cronológico de la ruta (RN23).

            '
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: El actor autenticado no es Técnico URP (inferido de `x-roles`).
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
  /proyectos/{idProyecto}/ficha-informacion-general:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    get:
      operationId: obtenerFichaInformacionGeneral
      summary: Consultar la Ficha de información general (Anexo A.3, FA-04)
      description: 'RN14: no editable, generada a partir de CU-PRE-01. RN15: 3 de sus campos (`objetivoProyecto`, `montoEstimadoInversion`, `descripcionProyecto`) se actualizan según la última Opinión Técnica emitida (CU-PRE-04, CU-PRE-17, CU-PRE-11 respectivamente) — antes de la primera Opinión Técnica, reflejan el valor original de CU-PRE-01 (`descripcionProyecto`/`montoEstimadoInversion`) o quedan vacíos (`objetivoProyecto`, que no existe como campo en CU-PRE-01; ver contrato-CU-PRE-03.5.md). RN17: `montoAjustadoEjecucion` solo se devuelve (no nulo) cuando el proyecto está en etapa de Ejecución.

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      - Coordinador SYMP
      responses:
        '200':
          description: Ficha de información general del proyecto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaInformacionGeneral'
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
  /proyectos/{idProyecto}/ficha-informacion-general/co-ejecutor:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    put:
      operationId: seleccionarCoEjecutor
      summary: Seleccionar la Unidad Ejecutora Co-ejecutora (botón radial "Co-ejecutor", RN16)
      description: 'RN16 es el único campo genuinamente editable de la Ficha de información general (a pesar de que esa pantalla también tiene un botón "Guardar" sujeto a RN19 — se interpreta que ese "Guardar" solo aplica a esta selección, ya que RN14 dice explícitamente que el resto de la ficha "será no editable"; ver contrato-CU-PRE-03.5.md). La activación del botón radial en sí (que solo habilita el listado de Unidades Ejecutoras en el cliente) no es una llamada separada: enviar este `PUT` ya representa la acción completa de activar+seleccionar+guardar.

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Coordinador SYMP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeleccionCoEjecutorRequest'
      responses:
        '200':
          description: Co-ejecutor asignado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaInformacionGeneral'
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: 'El actor autenticado no es Coordinador SYMP (RN16). ✅ Confirmado (04/09/2026, decisión funcional v1.3): el botón de selección radial "Co-ejecutor" está **oculto** (no solo deshabilitado) para cualquier actor distinto del Coordinador SYMP — resuelve la ambigüedad de interfaz que una versión previa de este fragmento dejaba sin resolver. Este 403 sigue siendo la restricción a nivel de API, ahora consistente con que el propio control ni siquiera se renderiza para otros roles.

            '
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto o la Unidad Ejecutora indicada no existen.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
  /proyectos/{idProyecto}/ficha-emergencia:
    parameters:
    - name: idProyecto
      in: path
      required: true
      schema:
        type: integer
        format: int64
    get:
      operationId: obtenerFichaEmergencia
      summary: Consultar la Ficha de proyectos de emergencia (Anexo A.4)
      description: 'Solo aplica a proyectos categorizados como "Proyecto de emergencia" (CU-PRE-01). Devuelve los campos no editables precargados desde CU-PRE-01 junto con los campos editables ya guardados (nulos si aún no se ha completado el formulario).

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      responses:
        '200':
          description: Ficha de proyectos de emergencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaEmergencia'
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: El actor autenticado no es Técnico URP (inferido de `x-roles`).
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe o no está categorizado como de emergencia.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
    put:
      operationId: registrarFichaEmergencia
      summary: Registrar la Ficha de proyectos de emergencia (botón "Guardar", FA-05)
      description: 'FA-05, paso 5.5: si todos los campos obligatorios están diligenciados, guarda la información y remite el proyecto a Viabilidad (CU-PRE-24) — efecto secundario modelado como transición de `Proyecto.estado` a `EN_VIABILIDAD` (ver CU-PRE-01.openapi.yaml). Si falta información, el Sistema no guarda nada y responde 400 con el mensaje "Existen campos sin diligenciar".

        '
      tags:
      - Preinversión - Selección y Registro de Etapas
      security:
      - bearerAuth: []
      x-roles:
      - Técnico URP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FichaEmergenciaRequest'
      responses:
        '200':
          description: Ficha registrada; proyecto remitido a Viabilidad (CU-PRE-24).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaEmergencia'
        '400':
          description: '"Existen campos sin diligenciar" (FA-05, paso 5.5): faltan uno o más de los campos obligatorios (Planteamiento del problema, Producto, Distrito, Población objetivo — Departamento se valida de forma transitiva a través de Distrito, ver contrato-CU-PRE-03.5.md). `Error.detalles` lista los campos faltantes.

            '
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '401':
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '403':
          description: El actor autenticado no es Técnico URP (inferido de `x-roles`).
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
        '404':
          description: El proyecto no existe o no está categorizado como de emergencia.
          content:
            application/json:
              schema:
                $ref: ../administracion/CU-ADM-03-manejo-errores.openapi.yaml#/components/schemas/Error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TipoCapital:
      type: string
      description: Catálogo de tipo de capital (Anexo C.1), listado fijo.
      enum:
      - CAPITAL_FISICO
      - CAPITAL_HUMANO
      - CAPITAL_INSTITUCIONAL
      - OTROS_CAPITALES
    TamanioProyecto:
      type: string
      description: Catálogo de tamaño del proyecto según monto (Anexo C.3), listado fijo.
      enum:
      - PEQUENIO
      - MEDIANO
      - GRANDE
    ComplejidadProyecto:
      type: string
      description: 'Catálogo de complejidad del proyecto (Anexo C.4), listado fijo. Incluye "TODAS_LAS_COMPLEJIDADES" porque el propio catálogo original lo lista como una cuarta opción seleccionable, no como una categoría derivada de las otras tres.

        '
      enum:
      - BAJA
      - MEDIA
      - ALTA
      - TODAS_LAS_COMPLEJIDADES
    NombreEtapa:
      type: string
      description: 'Etapas de la Ruta de Preinversión. No incluye un valor distinto para "Estudio General": RN07/RN08 tratan a Estudios Generales y Programa como usuarios de las mismas PERFIL/EJECUCIÓN que Proyecto, no como una etapa con nombre propio.

        '
      enum:
      - PERFIL
      - PREFACTIBILIDAD
      - FACTIBILIDAD
      - DISENO
      - EJECUCION
    CriteriosCalificacion:
      type: object
      description: Los 3 criterios calificables de la pantalla "Ruta de Preinversión" (RN01).
      required:
      - tipoCapital
      - tamanioProyecto
      - complejidad
      properties:
        tipoCapital:
          $ref: '#/components/schemas/TipoCapital'
        tamanioProyecto:
          $ref: '#/components/schemas/TamanioProyecto'
        complejidad:
          $ref: '#/components/schemas/ComplejidadProyecto'
    RutaPreinversionSugerida:
      type: object
      description: Respuesta de `generarRutaPreinversion` (no persistida).
      required:
      - criterios
      - etapasSugeridas
      properties:
        criterios:
          $ref: '#/components/schemas/CriteriosCalificacion'
        etapasSugeridas:
          type: array
          description: Según la combinación del Anexo B.2 (RN10).
          items:
            $ref: '#/components/schemas/NombreEtapa'
...
    RutaPreinversion:
      type: object
      description: Estado de la Ruta de Preinversión de un proyecto (Anexo A.2).
      required:
      - idProyecto
      - etapasAceptadas
      - fueModificada
      properties:
        idProyecto:
          type: integer
          format: int64
        criterios:
          type: object
          allOf:
          - $ref: '#/components/schemas/CriteriosCalificacion'
          nullable: true
          description: 'Nulo si el proyecto no es de iniciativa PROYECTO (RN07/RN08 no usan calificación de criterios), o si aún no se ha generado ninguna ruta.

            '
        etapasAceptadas:
          type: array
          description: Ruta actualmente vigente, reflejada en Registro de Etapas.
          items:
            $ref: '#/components/schemas/NombreEtapa'
        fueModificada:
          type: boolean
          description: true si la ruta vigente proviene de `modificarRutaPreinversion`, no de la sugerencia original.
        justificacionUltimaModificacion:
          type: string
          nullable: true
    Etapa:
      type: object
      description: Fila de la tabla "Registro de Etapas" (Anexo A.1).
      required:
      - nombreEtapa
      - habilitadoParaRegistro
      - tieneOpinionTecnica
      - bloqueadaPorModificacion
      properties:
        nombreEtapa:
          $ref: '#/components/schemas/NombreEtapa'
        costo:
          type: number
          format: double
          nullable: true
          description: 'RN04: obligatorio para guardar. RN05/RN11: para EJECUCIÓN, el servidor lo sobrescribe automáticamente (Presupuesto de inversión u Opinión Técnica); ver advertencia de RN22 en `actualizarEtapas`.

            '
        fechaInicio:
          type: string
          pattern: ^\d{2}/\d{2}/\d{4}$
          nullable: true
          description: '"Fecha estimada de inicio", formato dd/mm/aaaa — ✅ confirmado (04/09/2026, decisión funcional v1.3): el Anexo B.1 decía "MM/AA" y el mockup mostraba "mm/aaaa"; RN04 (dd/mm/aaaa) queda como el formato definitivo. Ya no se modela como texto enteramente libre. RN23: no puede ser posterior a `fechaFin` de la misma etapa, ni anterior a la `fechaFin` de la etapa previa de la ruta (cuando ambas tienen fechas completas).

            '
        fechaFin:
          type: string
          pattern: ^\d{2}/\d{2}/\d{4}$
          nullable: true
          description: Fecha estimada de finalización, formato dd/mm/aaaa. Mismo criterio que fechaInicio (RN04/RN23).
        habilitadoParaRegistro:
          type: boolean
          description: 'Si el botón de navegación de esta etapa hacia identificación/ formulación/evaluación/programación (CU-PRE-04 a CU-PRE-26) está habilitado. Pasa a true la primera vez que se guardan sus fechas (FA-01.1.2/FA-03.2); PERFIL y EJECUCIÓN lo tienen desde el inicio por defecto (RN09).

            '
        tieneOpinionTecnica:
          type: boolean
          description: 'Si ya se emitió Opinión Técnica para esta etapa (CU-PRE-26). Determina, junto con RN09, si la siguiente etapa de la ruta se habilita, y es la condición que dispara el bloqueo de RN13 al modificar la ruta.

            '
        bloqueadaPorModificacion:
          type: boolean
          description: 'RN13: true si esta etapa ya tenía Opinión Técnica y una modificación de ruta posterior requiere pasar nuevamente por ella. No se pierde su información, pero queda de solo lectura hasta obtener Opinión Técnica nuevamente.

            '
    EtapaRegistroRequest:
      type: object
      description: '`costo`/`fechaInicio`/`fechaFin` no son obligatorios a nivel de schema: RN19 (bordes rojos) es puramente visual, no bloquea el guardado — ver descripción de `actualizarEtapas`. `nombreEtapa` sí es obligatorio, ya que identifica la fila dentro del array.

        '
      required:
      - nombreEtapa
      properties:
        nombreEtapa:
          $ref: '#/components/schemas/NombreEtapa'
        costo:
          type: number
          format: double
        fechaInicio:
          type: string
          pattern: ^\d{2}/\d{2}/\d{4}$
          description: Formato dd/mm/aaaa (RN04, confirmado v1.3).
        fechaFin:
          type: string
          pattern: ^\d{2}/\d{2}/\d{4}$
          description: Formato dd/mm/aaaa (RN04, confirmado v1.3).
    ActualizarEtapasRequest:
      type: object
      required:
      - etapas
      properties:
        etapas:
          type: array
          items:
            $ref: '#/components/schemas/EtapaRegistroRequest'
...
    SeleccionCoEjecutorRequest:
      type: object
      required:
      - idUnidadEjecutoraCoEjecutora
      properties:
        idUnidadEjecutoraCoEjecutora:
          type: integer
          format: int64
    FichaInformacionGeneral:
      type: object
      description: 'Ficha no editable (RN14), salvo `coEjecutor` (RN16). No hereda de `Proyecto` (CU-PRE-01) vía `allOf`: sus campos son una selección curada y renombrada de RN14, no una copia estructural de `Proyecto`.

        '
      required:
      - idProyecto
      - institucion
      - unidadEjecutora
      - iniciativaInversion
      - nombreProyecto
      - montoEstimadoInversion
      - sector
      - ejeTematico
      - esProyectoGrdGrcAcc
      - esProyectoEmergencia
      - descripcionProyecto
      properties:
        idProyecto:
          type: integer
          format: int64
        institucion:
          $ref: ./CU-PRE-01.openapi.yaml#/components/schemas/InstitucionResumen
        unidadEjecutora:
          $ref: ./CU-PRE-01.openapi.yaml#/components/schemas/UnidadEjecutoraResumen
        coEjecutor:
          type: object
          allOf:
          - $ref: ./CU-PRE-01.openapi.yaml#/components/schemas/UnidadEjecutoraResumen
          nullable: true
          description: RN16 — asignado vía `seleccionarCoEjecutor` (Coordinador SYMP).
        iniciativaInversion:
          $ref: ./CU-PRE-01.openapi.yaml#/components/schemas/IniciativaInversion
        nombreProyecto:
          type: string
        montoEstimadoInversion:
          type: number
          format: double
          description: 'RN15: se actualiza según la última Opinión Técnica (campo "Total inversión" de CU-PRE-17); antes de la primera Opinión Técnica, refleja el valor original de CU-PRE-01.

            '
        montoAjustadoEjecucion:
          type: number
          format: double
          nullable: true
          description: RN17 — solo no nulo cuando el proyecto está en etapa de Ejecución. RN18 — se alimenta automáticamente.
        sector:
          allOf:
          - $ref: ../administracion/catalogo/CU-ADM-02-catalogos.openapi.yaml#/components/schemas/SectorResumen
          description: RN14 lista "Macrosector" como campo propio; se expone anidado en `sector.macrosector` (mismo objeto ya definido en CU-PRE-01), no como campo separado.
        ejeTematico:
          $ref: ../administracion/catalogo/CU-ADM-02-catalogos.openapi.yaml#/components/schemas/EjeTematicoResumen
        esProyectoGrdGrcAcc:
          type: boolean
          description: 'RN14 lo describe como un único campo booleano ("Es Proyecto GRD/GRC/ACC"), pero CU-PRE-01 modela GRD/GRC/ACC como 3 arrays de categorías independientes (medidasGrd/medidasGrc/medidasAcc). Se modela aquí como un valor derivado (true si alguno de esos 3 arrays no está vacío en `Proyecto`), no como un campo propio almacenado — ver contrato-CU-PRE-03.5.md. Nomenclatura: se usó el nombre de RN14 ("GRD/GRC/ACC"), no el del mockup del Anexo A.3 ("GRD/ACC", sin "GRC") — discrepancia ya señalada por el propio CU.

            '
        esProyectoEmergencia:
          type: boolean
        tipoEvento:
          type: string
          nullable: true
        numeroDecretoLegislativo:
          type: string
          nullable: true
        ejePlanGobierno:
          type: object
          allOf:
          - $ref: ../administracion/catalogo/CU-ADM-02-catalogos.openapi.yaml#/components/schemas/EjePlanGobiernoResumen
          nullable: true
        planSectorialRegional:
          type: object
          allOf:
          - $ref: ../administracion/catalogo/CU-ADM-02-catalogos.openapi.yaml#/components/schemas/PlanSectorialRegionalResumen
          nullable: true
        objetivoProyecto:
          type: string
          nullable: true
          description: 'RN15: procede del campo "Objetivo General" de CU-PRE-04 "Identificación" (no de CU-PRE-01, que no tiene un campo equivalente — confirmado al revisar Proyecto/ProyectoRequest en CU-PRE-01.openapi.yaml). Nulo hasta que CU-PRE-04 lo registre.

            '
        descripcionProyecto:
          type: string
          description: 'RN15: se actualiza según la última Opinión Técnica (tabla "Descripción Técnica" de CU-PRE-11); antes de eso, refleja el valor original de CU-PRE-01.

            '
...